+++ src/app/components/login/login.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { MatSnackBar } from '@angular/material';
 import { Router } from '@angular/router';
 
 import { CacheService } from '../../services/cache/cache.service';
@@ -16,7 +17,8 @@ export class LoginComponent {
   loginMessage: string;
   isLoading = false;
 
-  constructor(private router: Router, private messageService: MessageService, private cache: CacheService) {}
+  constructor(private messageService: MessageService, private cache: CacheService,
+              private router: Router, private snackBar: MatSnackBar) {}
 
   attemptLogin(): void {
     this.loginMessage = '';
@@ -41,7 +43,13 @@ export class LoginComponent {
       })
       .catch((error) => {
         this.isLoading = false;
-        this.loginMessage = error;
+        this.loginMessage = 'Incorrect username or password. Try again.';
+        if (error) {
+          this.snackBar.open('Encountered an error while trying to login.', 'Dismiss', {
+            duration: 30000,
+            panelClass: 'mat-snack-bar-error'
+          });
+        }
       });
   }
 }