+++ src/app/components/order-history/order-history.module.ts
@@ -9,7 +9,6 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { CacheService } from '../../services/cache/cache.service';
-import { MessageService } from '../../services/message/message.module';
 
 import { OrderHistoryComponent } from './order-history.component';
 
@@ -33,8 +32,7 @@ export { OrderHistoryComponent } from './order-history.component';
     MatTooltipModule
   ],
   providers: [
-    CacheService,
-    MessageService
+    CacheService
   ]
 })
 export class OrderHistoryModule { }