+++ src/app/services/message/message.service.ts
@@ -4,6 +4,7 @@ import { Subject } from 'rxjs/Subject';
 
 import { environment } from '../../../environments/environment';
 
+import { reject } from 'q';
 import { Batch } from '../../core/batch';
 import { DeliveryLocation } from '../../core/delivery-location';
 import { Edge } from '../../core/edge';
@@ -17,6 +18,8 @@ import { Vertex } from '../../core/vertex';
 
 @Injectable()
 export class MessageService {
+  private static maxNumberOfRetries = 1;
+  private static retryErrorCode = 418;
 
   private sessionKey: string;
   private user: UserInfo;
@@ -26,26 +29,40 @@ export class MessageService {
 
   constructor(private http: HttpClient) {}
 
-  getEnvironmentDetails(): Promise<EnvironmentDetails> {
+  getEnvironmentDetails(retryCount = 0): Promise<EnvironmentDetails> {
     const body = {
       username: this.user.username,
       sessionKey: this.sessionKey
     };
-    return this.http.post(environment.backendUrl + 'getEnvironmentDetails', body)
-      .toPromise().then((response: EnvironmentDetails) => response);
+    return this.http.post(environment.backendUrl + 'getEnvironmentDetails', body).toPromise()
+      .then((response: EnvironmentDetails) => response)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getEnvironmentDetails(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  login(username: string, password: string): Promise<UserInfo> {
+  login(username: string, password: string, retryCount = 0): Promise<UserInfo> {
     return this.http.post(environment.backendUrl + 'login', {username: username, password: password}).toPromise()
       .then((response: {sessionKey: string, userInfo: UserInfo}) => {
         this.sessionKey = response.sessionKey;
         this.user = response.userInfo;
         this.userUpdates.next(this.user);
         return this.user;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.login(username, password, retryCount + 1);
+        }
+        return Promise.reject(err.status === MessageService.retryErrorCode);
       });
   }
 
-  logout(): Promise<void> {
+  logout(retryCount = 0): Promise<void> {
     // todo: make sure all subscriptions are finished
     // todo: change flag in the service, subs check that flag before next request, once they all give the "OK" resolve to this promise
     const body = {username: this.user.username, sessionKey: this.sessionKey};
@@ -54,10 +71,17 @@ export class MessageService {
         this.sessionKey = undefined;
         this.user = undefined;
         this.userUpdates.next(undefined);
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.logout(retryCount + 1);
+        }
+        return Promise.reject(err);
       });
   }
 
-  updateAccountInfo(user: UserInfo): Promise<void> {
+  updateAccountInfo(user: UserInfo, retryCount = 0): Promise<void> {
     const body = {
       username: user.username,
       sessionKey: this.sessionKey,
@@ -67,90 +91,143 @@ export class MessageService {
       .then(() => {
         this.user = user;
         this.userUpdates.next(this.user);
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.updateAccountInfo(user, retryCount + 1);
+        }
+        return Promise.reject(err);
       });
   }
 
-  sendBatch(batch: Batch): Promise<void> {
+  sendBatch(batch: Batch, retryCount = 0): Promise<void> {
     const body = {
       username: this.user.username,
       sessionKey: this.sessionKey,
       batchId: batch.id
     };
-    return this.http.post(environment.backendUrl + 'sendBatch', body).toPromise().then(() => null);
-  }
-
-  getIncomingBatches(): Promise<Batch[]> {
-    const body = {username: this.user.username, sessionKey: this.sessionKey};
-    return this.http.post(environment.backendUrl + 'getIncomingBatches', body)
-      .toPromise().then((response: {batches: Batch[]}) => {
-        return response.batches;
+    return this.http.post(environment.backendUrl + 'sendBatch', body).toPromise()
+      .then(() => null)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.sendBatch(batch, retryCount + 1);
+        }
+        return Promise.reject(err);
       });
   }
 
-  getOrderHistory(): Promise<Order[]> {
+  getIncomingBatches(retryCount = 0): Promise<Batch[]> {
     const body = {username: this.user.username, sessionKey: this.sessionKey};
-    return this.http.post(environment.backendUrl + 'getOrderHistory', body)
-      .toPromise().then((response: {orderHistory: Order[]}) => response.orderHistory);
+    return this.http.post(environment.backendUrl + 'getIncomingBatches', body).toPromise()
+      .then((response: {batches: Batch[]}) => {
+        return response.batches;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getIncomingBatches(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getOrderUpdates(ordersToMonitor: Order[]): Promise<Order[]> {
+  getOrderHistory(retryCount = 0): Promise<Order[]> {
+    const body = {username: this.user.username, sessionKey: this.sessionKey};
+    return this.http.post(environment.backendUrl + 'getOrderHistory', body).toPromise()
+      .then((response: {orderHistory: Order[]}) => response.orderHistory)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getOrderHistory(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
+  }
+
+  getOrderUpdates(ordersToMonitor: Order[], retryCount = 0): Promise<Order[]> {
     const body = {
       username: this.user.username,
       sessionKey: this.sessionKey,
       orders: ordersToMonitor
     };
-    return this.http.post(environment.backendUrl + 'getOrderUpdates', body)
-      .toPromise().then((response: {orders: Order[]}) => response.orders);
+    return this.http.post(environment.backendUrl + 'getOrderUpdates', body).toPromise()
+      .then((response: {orders: Order[]}) => response.orders)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getOrderUpdates(ordersToMonitor, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  placeOrder(orderInfo: OrderInfo, deliveryLocation: DeliveryLocation): Promise<Order> {
+  placeOrder(orderInfo: OrderInfo, deliveryLocation: DeliveryLocation, retryCount: number = 0): Promise<Order> {
     const body = {
       username: this.user.username,
       sessionKey: this.sessionKey,
       orderInfo: orderInfo,
       deliveryLocation: deliveryLocation
     };
-    return this.http.post(environment.backendUrl + 'placeOrder', body)
-      .toPromise().then((response: {order: Order}) => {
+    return this.http.post(environment.backendUrl + 'placeOrder', body).toPromise()
+      .then((response: {order: Order}) => {
         this.orderPlacedUpdate.next(response.order);
         return response.order;
+      }).catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.placeOrder(orderInfo, deliveryLocation, retryCount + 1);
+        }
+        return Promise.reject(err);
       });
   }
 
-  getSystemDetails(): Promise<SystemDetails> {
+  getSystemDetails(retryCount = 0): Promise<SystemDetails> {
     const body = {
       username: this.user.username,
       sessionKey: this.sessionKey
     };
-    return this.http.post(environment.backendUrl + 'getSystemDetails', body)
-      .toPromise().then((response: SystemDetails) => response);
-  }
-
-  getMap(): Promise<LocationMap> {
-    return this.http.get(environment.backendUrl + 'getMap')
-      .toPromise().then((response: any) => {
-        const map = new LocationMap();
-
-        response.vertices.forEach((vertex: any) => {
-          map.vertices.push(new Vertex(vertex));
-        });
-
-        response.edges.forEach((edge: any) => {
-          map.edges.push(new Edge(edge));
-        });
-
-        return map;
+    return this.http.post(environment.backendUrl + 'getSystemDetails', body).toPromise()
+      .then((response: SystemDetails) => response)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getSystemDetails(retryCount + 1);
+        }
+        return Promise.reject(err);
       });
   }
 
-  getPath(vertices: Edge[]): Promise<Edge[]> {
+  getMap(retryCount = 0): Promise<LocationMap> {
+    return this.http.get(environment.backendUrl + 'getMap').toPromise()
+      .then((response: any) => {
+        const map = new LocationMap();
+        response.vertices.forEach((vertex: any) => {
+          map.vertices.push(new Vertex(vertex));
+        });
+        response.edges.forEach((edge: any) => {
+          map.edges.push(new Edge(edge));
+        });
+        return map;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getMap(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
+  }
+
+  getPath(vertices: Vertex[], retryCount = 0): Promise<Edge[]> {
     const body = {
       username: '',
       sessionKey: '',
       vertexValues: vertices
     };
-    return this.http.post(environment.backendUrl + 'getPath', body)
-      .toPromise().then(((response: any[]) => {
+    return this.http.post(environment.backendUrl + 'getPath', body).toPromise()
+      .then(((response: any[]) => {
         const convertedEdges: Edge[] = [];
 
         response.forEach(edge => {
@@ -158,7 +235,14 @@ export class MessageService {
         });
 
         return convertedEdges;
-      }));
+      }))
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getPath(vertices, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
 }